requirements.txt
Django==4.1.6
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-jwt